Version function to avoid potential race (#832)
Function code and function metadata are implemented in two separate transactions - once through dynamodb, and once through the lambda commit process. Both of these can complete unpredictably, and as a result it's possible for a heavily used function to be executed using a committed version of the metadata in dynamodb that's aligned with a different version of the code.

This PR resolves this through adding the version of the code to the function metadata, and then under execution executing only the version specified in the metadata itself. Old versions are kept around for a fixed interval before being garbage collected.