251. Refactor session handling to streamline use (#1066)
This PR contains a tactical rethink on how the session handling process works, largely on how sessions are committed and the necessary calls to perform afterwards.

### Change to operation polling

The process now looks as follows (this is highlighted for instance creation through sessions, but is true for all cases where operations are used including connector and integration creation):
1. POST /integration/id/session/sid
```
{
  operationId: '282ab0e9-8f78-46c2-b5f1-b3b6e19a6c8b',
  target: 'http://localhost:3001/v2/a/s/i/int-123/instance?operation=282ab0e9-8f78-46c2-b5f1-b3b6e19a6c8b',
  statusOnly: 'http://localhost:3001/v2/a/s/operation/282ab0e9-8f78-46c2-b5f1-b3b6e19a6c8b'
}
```
2. GET `http://localhost:3001/v2/a/s/i/int-123/instance?operation=282ab0e9-8f78-46c2-b5f1-b3b6e19a6c8b`
```
202 HTTP/1.1
```
...
```
200 HTTP/1.1

{
      "total": 1,
      "items": [
        {
          "id": "f3f9ed27-7b89-45d2-b3da-7f102cb04112",
          "data": {
            "conn": {
              "tags": {
                "session.master": "06576e70-1c89-4d60-a9a5-9f2c0be76e64"
              },
              "entityId": "daa93e23-dde7-4720-96a9-9d69f9f9fc68",
              "accountId": "acc-12d136912f0c4912",
              "entityType": "identity",
              "parentEntityId": "test-session-206-301251475-con",
              "subscriptionId": "sub-a447d98de09c4cfe",
              "parentEntityType": "connector"
            },
            "form": {
              "newData": "for instance"
            }
          },
          "tags": {
            "session.master": "06576e70-1c89-4d60-a9a5-9f2c0be76e64"
          },
          "version": "82b1e0bc-e0b3-4b17-b305-fad852745523"
        }
      ]
    }
```
Similar results for `/integration?operation=...` and `/connector?operation=...` endpoints are also available.

The `statusOnly` url is there to be passed on to the untrusted frontend for driving progress bars or spinners, whereas the result of a GET operation on the `target` url is output identical to that of a search operation on the same endpoint.  All three search endpoints - integrations, connectors, and instances - support an `operation` parameter for getting the resulting integration, connector, or instance respectively.

### Removal of the `/session/:id/result` call

The /session/id/result call is removed completely, with the `output` field being returned on a `GET /session/sid` operation.  This was originally segregated for security reasons, but because the `GET /session/sid` was already a permissioned operation, the value diminished. Having a separate call didn't seem as warranted, though the author welcomes additional perspectives.

The primary use of this was in the test code, which has been replaced with a direct database lookup for similar levels of validation.

Uses from within an integration will be satisfied by the `GET /session/:id` call providing the output parameters of a previous operation, if the permissions allow.

### Rename of operation.code to operation.statusCode

It seemed to be a much more consistent naming convention, so the opportunity was taken to remove the outlier.

### Change a delete_function TooManyRequestsException to 429

Somehow it was a 428 error code, almost certainly due to clumsy fingers by some unnamed fool.