ENG-190 Segment Integration (#1069)
Adds support for Segment as another destination (along with ElasticSearch) of analytic data. The process of offloading this data is, just like with ElasticSearch, handled by the Lambda Analytics function we have.

To update/create a deployment with Segment, you can use the `--segmentKey` of the `fuse-ops deployment add` command.