ENG-453 Refreshing caches during tests
The following tests were failing on the CI/CD pipeline:

```
 FAIL  test/v1/function.staticip.test.ts (103.853s)
  - Subscription with staticIp=true › PUT supports setting staticIP=true
  - Subscription with staticIp=true › PUT multiple times on the same function
  - Subscription with staticIp=true › PUT with new compute values updates compute and computeSerialized
```

The problem was that we had a subscription cache refresh call that was not refreshing all instances. The intention of this PR is to work around this problem by submitting a bunch of requests at the same time. This worked with a 4 instances deployment, but there are chances that this might randomly fail, eventually.