ENG-552 - Move session commit to async model (#1178)
This PR changes the session commit process to an async model. Now, when the client issues a POST request to .../:sessionId/commit, the response returns a status code of 202 and a targetUrl of the entity being created. At the beginning the operation is creating and the status is processing, then eventually the status changes to success if everything goes well.